app: make 'thinking' banner work; cpu responsive
Waahahaha I'm a madman 🤪

Actually, I'm using setTimeout(f, t) to run f after t milliseconds; this
enables me to make some DOM changes and, crucially, return from the
event handler (so that they are updated on screen).

Only *then* does f run, which in this case undoes some of the DOM
changes, in addition to actually calling the longer-running task of
getting and making the CPU's move.

Now, there is also code that has to run immediately after the CPU's
move... so pass in a callback (more properly called a continuation in
this case) and then run that sucker!