app: really focus cycleMovesBtn after move played
We have to workaround the keyboard/gamepad handlers, which maintain
internal state about which node is focused (selectableItems,
selectItemsIdx).

This time, we set the focus only once input has been transferred to the
user; and we also update the internal state to be correct!

A good rework would have probably two or three functions:
- set_selectable([{node, type}...]), which builds the internal state of
  selectables
- set_focus(?), which allows us to manually change the focus
  (correctly updates the internal state)

And it would establish the event handlers in a way that they depend on
the passed selectables (can dispatch on type to decide what to do
correctly).